requirement.txt
absl-py==0.7.1
astor==0.8.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
bokeh==1.3.2
certifi==2019.6.16
cffi==1.14.5
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
configparser==5.0.1
cycler==0.10.0
Cython==0.29.13
dataclasses==0.8
decorator==4.4.0
defusedxml==0.6.0
dlib==19.17.0
docker-pycreds==0.4.0
entrypoints==0.3
future==0.18.2
gast==0.2.2
gitdb==4.0.5
GitPython==3.1.11
graphviz==0.11.1
grpcio==1.22.0
h5py==2.9.0
horovod==0.21.3
idna==2.8
imageio==2.9.0
intel-openmp==2020.0.133
ipykernel==5.1.1
ipython==7.7.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.14.1
Jinja2==2.10.1
joblib==0.13.2
json5==0.8.5
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
jupyterlab==1.0.4
jupyterlab-server==1.0.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
mkl==2019.0
mock==3.0.5
nbconvert==5.5.0
nbformat==4.4.0
networkx==2.3
notebook==6.0.0
numpy==1.16.0
opencv-python==4.1.0.25
packaging==19.1
pandas==0.25.0
pandocfilters==1.4.2
parso==0.5.1
path==15.1.2
pathtools==0.1.2
patsy==0.5.1
Pebble==4.6.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
plotly==4.0.0
prometheus-client==0.7.1
promise==2.3
prompt-toolkit==2.0.9
protobuf==3.13.0
psutil==5.7.2
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.4.2
pyparsing==2.4.2
pyrsistent==0.15.4
python-apt==1.6.5+ubuntu0.3
python-dateutil==2.8.0
pytz==2019.2
PyWavelets==1.1.1
PyYAML==5.1.2
pyzmq==18.0.2
qtconsole==4.5.2
requests==2.22.0
retrying==1.3.3
scikit-image==0.17.2
scikit-learn==0.21.3
scipy==1.3.0
seaborn==0.9.0
Send2Trash==1.5.0
sentry-sdk==0.19.1
shortuuid==1.0.1
six==1.15.0
smmap==3.0.4
statsmodels==0.10.1
subprocess32==3.5.4
tensorboard==1.13.1
tensorboard-logger==0.1.0
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tifffile==2020.9.3
torch==1.6.0+cu101
torchsummaryX==1.3.0
torchvision==0.7.0+cu101
tornado==6.0.3
tqdm==4.50.2
traitlets==4.3.2
urllib3==1.25.3
wandb==0.10.8
watchdog==0.10.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.5
widgetsnbextension==3.5.1
yacs==0.1.8
